hit detection with goal.
Changed goal position in golf map to vec2, and assigned it's value in GolfMap.cc.

Added a vec2 pointer in golfball so it can check the position itself.

Added switch map function in the app, this will handle logic for switching map and will be called by the ball when appropriate.

Added hit detection for goal in ball, and a method that is called when it occurs.