Save and read
Reads highscores from file.

Save format updated so that numbers are also character's , the number is the character's ascii index. This allows for scores as high as 127 using only one character to save it.

I've also done some more cleanup in app. Along with adding more necessary variables in the app. Along with a highscores struct.